requirements.txt
flet
pytesseract
python-dotenv
opencv-python
pypdf
